Added login request to PostInfo
