+++ src/index.ios.js
@@ -182,7 +182,8 @@ export default class Scribe extends Component {
   }
 
   login() {
-    lock.show({ closable: true }, (err, profile, token) => {
+    console.log('loggin in from index');
+    lock.show({ zIndex:10, closable: true }, (err, profile, token) => {
       if (err) {
         console.log(err);
         return;
